Merged PR 50731: Set default orders mega batch size to 4000
## What's being changed

The default mega batch size for orders sync is being reduced to 4000.

## Why it's being changed

The mega batch size (in bulk syncs) is the max number of records that can be included in a JSON blob queued with the importer. We've had some reports from merchants that at 5000 records order sync cannot save complete data into `import_data` because it exceeds 16MB. This then results in failed imports with the message 'Error unknown' (which in this case means the serializer could not json_decode the import_data string to valid JSON).

## How to review / test this change

- Remove any `mega_batch_size_orders` keys from `core_config_data` and clear cache
- Reset orders
- Run order sync

Related work items: #239213