Merged PR 50917: Add nyholm/psr7 to dependencies
## What's being changed

We've added a specific PSR-17 implementation to the module to support use of our SDK in 2.3.7.

## Why it's being changed

We were unable to instantiate a V3 client in 2.3.7. We saw this in the logs:
```
[2024-01-24 20:20:06] report.CRITICAL: Type Error occurred when creating object: Dotdigital\V3\Client, Too few arguments to function Laminas\Diactoros\ResponseFactory::__construct(), 0 passed in /Users/alastairmucklow/Sites/magento-cc-2.3.7/vendor/php-http/discovery/src/ClassDiscovery.php on line 219 and at least 1 expected [] []
```

We were relying on Magento to provide the PSR-17 factory implementation, but in the case of older versions of laminas/laminas-diactoros, the package did not provide a ResponseFactory. In this case, Magento was auto-generating its own ResponseFactory, but this required ObjectManager to be passed as the first argument.

nyholm/psr7 provides its own factory thus resolves this muddle.

## How to review / test this change

- Run:
```
composer config extra.discovery.psr/http-factory-implementation Nyholm\\Psr7\\Factory\\Psr17Factory
```
- Run composer update (add `"nyholm/psr7": "^1.8"` to your 2.3.7 composer.json)
- In 2.3.7, log in to the admin
- Save Accounts > Dotdigital
- See no error

## Notes

We already include a PSR-18 client implementation in our module dependencies (i.e. we require `"php-http/curl-client"`).

Related work items: #240564