Merged PR 50975: Add Updated At column to report table
## What's being changed

In this PR, we are making changes to the `dotdigitalgroup_queue_grid.xml` file.

The `dotdigitalgroup_queue_grid.xml` file has been updated to include a new column in the grid for `updated_at`. This column will provide information about when each message in the queue was last updated, making it easier for users to understand the status and content of each message.

## Why it's being changed

The change is being made to improve the visibility and manageability of the queue messages. By providing the `updated_at` information in the grid, we can make it easier for users to monitor and manage the queue.

## How to review / test this change

- Pull the changes from this PR into your local environment.
- Run the Magento `cache:flush` command to apply the changes.
- Navigate to the Dotdigitalgroup queue grid in the Magento admin panel. You should see the new `updated_at` column in the grid.
- Test the filtering and sorting functionality of the new `updated_at` column. Apply a filter to the column and verify that the grid is updated to only show messages that match the filter. Apply a sort order to the column and verify that the messages are sorted correctly.

Related work items: #240805