Merged PR 51635: Update templates and HTML in customer account
## What's being changed

Various bits of templating and HTML in our customer account view (connector/customer/index).

## Why it's being changed

The main goal here was to make it easier to share templates between Luma and Hyvä themes. With a few small adjustments, we don't need to duplicate code in this area in our compatibility module for Hyvä.

The logic in the method `shouldRedirectToConnectorCustomerIndex` has changed so that merchants _only_ need to have enabled an account connection to load our layout (they don't need to be displaying data fields, lists or preferences to customers as well). This allows merchants to turn on consent for Email, and see custom consent text in the customer account, without the need to switch other components on (we discovered this limitation by accident).

The menu item for 'Newsletter Subscriptions' now uses a `SortLinkInterface` block so it appears in the correct place, and can be moved into a different sort order by merchants if they prefer. Because of the logic changes above, we can now use an `ifconfig` to control the appearance of this instead of needing our own custom block.

Various kludgy inline styles have been removed. There is a new module.less file for the only CSS rule worth preserving. We now use a native browser date picker instead of mage/calendar (which isn't ported to Hyvä).

## How to review / test this change

- Set up a Magento with 2 websites, one on Hyva theme, one on Luma theme
- Turn off Configuration > show lists, show data fields, show preferences
- If DD connection is enabled, connector/customer/index should load in the customer account
- Check the 'Newsletter Subscriptions' link in the menu, it should load in the correct position (above Sign Out in Hyvä)
- Turn on Configuration > show lists, show data fields, show preferences
- Check the display of nested and non-nested preferences
- Check the display of all data field types (Text, Date, Boolean, Number)
- Save and retrieve data field values, confirm DD is updated
- Run `setup:upgrade` and confirm the new CSS is applied

## Notes

Native browser date picker is [supported in all current browsers](https://caniuse.com/?search=date%20picker) except for Opera Mini.

Related work items: #239914