Merged PR 52058: Remove redundant keys from CartInsight data creation via V3
## What's being changed

This PR modifies the cart insight payload to remove the key and contact identifier.

## Why it's being changed

Previously when we posted this data via the V2 API, this data structure was required. The data inside the `json` node was used for the visible record, while the key and contact identifier were used by the platform when creating the record and attaching it to a contact. However in V3, this extra data is not required (there are additional params in the request) and this unmodified structure actually breaks the abandoned cart block in email templates.

## How to review / test this change

- Set up a program for abandoned cart in DD.
- Have the program trigger an email campaign that contains the Abandoned Cart block.
- Test abandoned cart sync. Confirm cart insight reaches DD.
- Confirm you receive an email containing your cart details.
- Test the order complete queue consumer (it should sent cart insight with an ORDER_COMPLETE cart phase).
- Test dummy data creation (click save in account creds to run).

Related work items: #245029