Merged PR 52320: Update quote email at guest checkout on change
## What's being changed

We have updated the conditionals for the checkout AJAX call to update the quote email.

## Why it's being changed

At guest checkout, email capture will update the quote email address as soon as an email address is entered. However, once an email is set, it can't be overridden with a revised email address. This means that incorrect email addresses can remain permanently associated with a quote and cannot be amended.

## How to review / test this change

- Add a breakpoint to: `app/code/Dotdigitalgroup/Email/Controller/Ajax/Emailcapture.php` execute method
- Checkout, input email into the email field.
- Observe the quote email update via the breakpoint
- On the confirmation screen, click to update the shipping details and change the email address, observe the quote email update.
- Observe a green build for integration tests

Related work items: #246360