Merged PR 52547: Update pref centre MFTF test
## What's being changed

One failing MFTF test.

## Why it's being changed

Logic governing the visibility of our customer preference centre layout connector/customer/index has changed in 4.25.0. Merchants only need to have enabled an account connection to load our layout (they don't need to be displaying data fields, lists or preferences to customers as well). This allows merchants to turn on consent for Email, and see custom consent text in the customer account, without the need to switch other components on.

As a result of this changed in behaviour, an MFTF test needed to change too.

## How to review / test this change

- Observe this test passing in the suite: `ApiEnabledAndNoPrefEnabledNoRedirectToPrefCenterTest`

Related work items: #247271