Merged PR 52928: Surface non-curl errors in REST client execute method
## What's being changed

The logic run during account validation, and in the `execute` method of our bundled REST client. I've also removed a superfluous log line in account validation.

## Why it's being changed

If the response from `curl_exec()` is null for any API call via this client, we check for a curl error and message accordingly. However we have encountered edge cases where the response is empty, but there was no curl error.

An example of where this loophole has a negative impact is when validating account credentials and calling /v2/account-info. Currently, in the no-response-but-not-curl-error scenario, the merchant sees a partial success in the admin.

## How to review / test this change

- Go to your Dotdigital account screen in the Magento admin
- Turn off wifi
- Set `$this->curlError` to null at the start of the `Rest::getCurlError()` method
- Click save config
- Confirm the account shows the same generic 'Authorization has been denied' message
- Confirm the logs indicate a 'No response' message
- Remove `$this->curlError = null` and retry
- Confirm the logs indicate a 'CURL ERROR' message
- Turn wifi back on and confirm validation succeeds

Related work items: #248997