Merged PR 54135: Fix merging of keyed arrays in order sync batch merge
## What's being changed

We're fixing the syntax of `mergeBatch()` in the order exporter.

## Why it's being changed

The affected code does not (always) preserve keys. Consider:
```
$megaBatch = [
    '1' => [
        '123' => $array123,
        '456' => $array456
    ]
];

$batch = [
    '1' => [
        '789' => $array789
    ]
];

mergeBatch($megaBatch, $batch);
```

In my system the 789 array is merged with the key '789' - but this shouldn't work. All my other local tests on arbitrary PHP arrays do not behave like this - until, that is, the syntax is changed to match this update.

## How to review / test this change

- Set your connector_developer_settings/import_settings/transactional_data to e.g. 5
- Set your connector_developer_settings/import_settings/mega_batch_size_orders to e.g. 10
- Have 10+ orders in your local Magento
- Run order sync
- Observe orders being batched in sets of 10, with the correct increment id keys

Related work items: #254262