Merged PR 54830: Product created_date falls back to old date if missing
## What's being changed

The product schema for catalog sync. Products without a `created_at` value will now have the `created_date` of '1970-01-01 01:00:00' (not '1970-01-01 00:00:00' because that ends up returning the current time, it must be after the start of the Unix epoch).

## Why it's being changed

Somehow it is possible for catalog products to have no `created_at` value.

## How to review / test this change

- Erase the created_at value for a product row in catalog_product_entity (or replace `$product->getCreatedAt()` with null in the code)
- Now try to sync the matching product
- Import data should show 1970-01-01 as the created_date
- Data should be successfully imported into DD

## Notes

As an alternative we could simply unset created_date if not supplied, but my understanding was this value was important for some functionality platform-side.

Related work items: #257276