Merged PR 54827: Pre-send order insight data for order automations
## What's being changed

The Automation Processor.

## Why it's being changed

Just as we pre-send contact data in advance of enrolment, so we should also pre-send order data for automations that require it. If a merchant is enrolling from order-related automations (new order, first order, order complete etc.) they may have steps in the target program that reference order properties. Now that automations are sent in near-time, this is often too fast for the sync (still running on a 15-minute cron).

## How to review / test this change

- Turn off cron (and therefore queues)
- Set up some order-type automations, plus some non-order-type automations
- Place some orders
- Run `bin/magento dotdigital:sync Automation`
- Confirm that order insight is sent successfully for your order-type automations

## Notes

The order pre-send will not:
- mark an order as imported
- reset ordered products

For now at least, order sync will re-export and re-send any pre-sent order data, in bulk, and will continue to do all the work of updating database rows, again in bulk.

Related work items: #252584