Merged PR 55371: Do not migrate orders for matching customer ids / website
## What's being changed

The InsertEmailContactTableCustomerSales migration type.

## Why it's being changed

This change prevents the migration of historical orders if we already have rows in `email_contact` with the same customer_id and website_id combination.

Ordinarily in Magento, if a customer with prior order history changes email address, their past orders are updated to keep the association. However we have come across installations where the order history has not been updated, leading to multiple rows in our table for the same customer in the same website.

## How to review / test this change

- Stay on develop branch
- Place an order as a customer
- Change your customer email address
- Go into sales_order and manually modify one of your updated order rows to have your previous email address
- Turn on Customer Configuration > Account Sharing > Global
- Run `dotdigital:migrate`
- See rows added for your new and your old email
- Switch to this branch
- Do not see any row added for your previous email

Related work items: #259692