Merged PR 55236: Allow product notifications to be configured differently in child scopes
## What's being changed

We have removed conditions from the backend model for the product notification config model, and we have added the `account_id` to the system config.

## Why it's being changed

The product notification `account_id` value was not being updated for multi-scopes due to the set condition being too strict. If a product notification was set at default level, this could not be overridden at website or store view level (because `getStoredAccountId` returned the inherited value from the default parent).

We have also allowed the value to be set manually via the CLI using the config set command for completion.

## How to review / test this change

- Run the following command: `bin/magento config:set "connector_api_credentials/api/account_id" 1234567890`
- Navigate to Configuration > Automation > Product Notifications
- Disable the feature
- Re-enable the feature and ensure the value in the config table matches your scoped account. You can run `bin/magento config:show "connector_api_credentials/api/account_id"` to confirm from the CLI.

Related work items: #257022