Merged PR 55616: Revert "No response in REST client" update
## What's being changed

We're reverting / amending [this commit](https://github.com/dotmailer/dotmailer-magento2-extension/commit/669e9aecd46a3353ac6cedff1195a48367407b02).

## Why it's being changed

In 4.25.2 we released a bugfix that was supposed to prevent a bad API response looking like a partial success, when connecting to a Dotdigital account. However, this change means that a number of V2 client methods that might return an empty array with a 200, (like `getContactsSuppressedSinceDate`) or a 204 No Content (like `deleteAddressBookContact`), are now dumping "No response in REST client" into the logs.

I've added an extra check into the account validation code to return false if the response is empty.

## How to review / test this change

- Go to your Dotdigital account screen in the Magento admin
- Set `$accountInfo` to null above L90 in Dotdigitalgroup\Email\Model\Apiconnector\Test
- Click to Save Config
- Confirm the account shows the expected 'Authorization has been denied' message

Related work items: #259561