Merged PR 55784: System messages to surface message queue problems
## What's being changed

This PR adds a new type of monitor to check problems with Dotdigital message queue consumers. The monitor runs alongside our other monitors (for cron, automations etc.) and will flag:
- any ddg messages marked with the status 'error'
- any ddg messages still with the status 'new' 1 hour after creation

Any problems are surfaced in a system message in the usual way, and/or emailed as a notification.

We also updated the dashboard to indicate when the last run of the consumers_runner cron was.

## Why it's being changed

With our transition to queues for faster sync times between Magento and Dotdigital, ensuring queue messages are consumed as expected and on time is now critical. We've noticed that the proper running of `consumers_runner` cron is not a given on merchant systems, and also that occasionally merchants have restricted running consumers to a subset (presumably to save on resource utilization), meaning that our queue consumers aren't running by default.

## How to review / test this change

- Turn on system messages and email alerts
- Turn off cron
- Queue some messages of different types, which will stay in pending
- Wait an hour or manually amend the `updated_at` value of your new queue messages
- Run the monitor task (bin/magento dotdigital:task Monitor)
- See a system message for Message Queue with the appropriate jobs highlighted
- Manually set some other messages to have status 6 (error)
- Re-run the monitor cron, see the system message updated to include errors
- Confirm that 'new' messages added in the last hour are not flagged
- Confirm that only ddg messages are flagged
- Confirm that email notifications work as expected (check the formatting and text of the email)
- Turn on cron, process your outstanding messages or otherwise resolve your bad statuses
- See no system message

## Notes
A reminder on how the notifications work.
- You can tune what the interval is. By default we find any errors or stuck-pending messages in the last 24 hours.
- The monitor cron runs every 15 mins, and updates the system message.
- Once the email notification kicks out, we won't send another one for another 24 hours (or whatever the interval is).

Related work items: #258685