Merged PR 56874: Do not sync Email, EmailType or OptInType as data fields
## What's being changed

This PR fixes handling of some data points previously treated as data fields in the context of CSV sync.

All contact exports now do not export Email or EmailType. Subscriber exports now set OptInType per subscriber depending on Customers > Newsletter > Subscription Options > Need to Confirm at _store level_ (this is a change - we used to check this at website level but the config is store-scoped).

## Why it's being changed

Email, EmailType and OptInType are not data fields - in V3, EmailType and OptInType are channel properties, and in fact even in V2, these data points are sent separately from data fields in API requests. It's only because of how we used to send these via CSV that they were handled the way they were.

## How to review / test this change

- Run customer sync, guest sync and subscriber sync
- All exported JSON should have `,"channelProperties":{"email":{"emailType":"html"}}` for each contact
- Subscriber exports should _not_ contain `optInType` yet
- Turn on Customers > Newsletter > Subscription Options > Need to Confirm in a store scope
- Re-sync subscribers
- Subscribers in the affected store scope should be synced with `"channelProperties":{"email":{"emailType":"html","optInType":"double"}}`
- Email, EmailType and OptInType should not be seen in data fields at all

Related work items: #264085, #265397, #265398