Merged PR 58003: Use constants for queue topic names
## What's being changed

This PR introduces class constants for all topic names. I've placed the constants in the same class as the `publish()` call (if they are only used in that one place) or else in a related model class.

## Why it's being changed

This is designed to protect us against error when updating queue topic names in future.

## How to review / test this change

- Test all the topic names, see each one output into the queue table (for subscription, pick one trigger that will publish a queue message)
- Confirm that we only use constants when calling `PublisherInterface::publish()`

Related work items: #265572