Merged PR 57124: Pre-sync subscribers via API V3
## What's being changed

SingleSubscriberSyncer and the DataFieldCollector. Subscribers being synced individually will now be sent via API V3, with specific exceptions logged with full details.

## Why it's being changed

Now that the SubscriberExporter and SubscriberWithSalesExporter return an array of SdkContacts, with channel properties already set, we can check the OptInType property on the hydrated model, without needing to look it up again prior to the Dotdigital request.

## How to review / test this change

- Switch off cron
- Test adding a new subscriber
- Start the `ddg.newsletter.subscription` queue consumer
- Check all expected data fields are present in Dotdigital
- Confirm that if Need to Confirm is set to Yes, optInType is captured as a channel property and sent with the request
- Repeat with a new subscriber automation (starting the `ddg.sync.automation` consumer)

Related work items: #266108