Merged PR 58402: New integration test for the OrderSaveAfter observer
## What's being changed

This PR adds an integration test for the OrderSaveAfter observer, covering off all its main functions. The old Automation\FirstOrderTest is removed.

## Why it's being changed

This is a critical observer that we ought to have test coverage of. A recent bug relating to a bad queue topic name has made this pressing.

Note that it wasn't possible to use Magento data fixtures in most of the tests. This is because a fixture like Magento/Sales/_files/order.php will create an order, triggering the observer, _before_ the test's setUp method has run, and too late to set config values. Hence we've reproduced two data fixtures as private methods.

FirstOrderTest is removed because it never actually passed the isEnabled check in the observer, hence didn't create an automation. It probably worked before because there was already an automation row left over after another test.

## How to review / test this change

- Observe a green build for integration tests

## Notes

For discussion - should we be saving an order in our table if the connector is not enabled?

Related work items: #264633