Merged PR 58360: Use UI component in favour of hashes for dmpt
## What's being changed

The tracking_code.phtml view file. Instead of loading an inline script, the template now loads a Magento UI component.

## Why it's being changed

Using UI components in favour of inline scripts is Magento's preferred approach. Now that CSP is set to restrict mode on payment pages and checkout, the data layer script tag is being blocked.

We previously used a hash method to whitelist the inline script. However this proved to be inconsistent, with some merchants reporting that dmpt.js was still blocked.

## How to review / test this change

- Use a regular Luma theme website
- Add a product to cart and proceed to checkout
- Confirm that the tracking script loads
- Confirm that requests are sent to https://r1.trackedweb.net/pagevisit
- Append a dm_i query string to your page URL like ?dm_i=0,0,0,0,0
- Confirm requests are sent to https://r1-t.trackedlink.net/PageTrack.ashx

## Notes

This change may have an impact on Hyvä theme compatibility.

Related work items: #271760