Merged PR 58292: Fix invalid datetime in consent sync
## What's being changed

The consent model. There is a new `getConsentDateTime()` method that overloads the default magic getter and returns a formatted date.

## Why it's being changed

We need to send ISO8601 date time values to Dotdigital. This worked before, it's a side-effect of the new strict API validation that came in last month. We see this in the logs:

```
dotdigital.DEBUG: Error when pushing consent batch: contacts:invalidValue - The value provided for 'contacts[0].consentRecords[0].dateTimeConsented' is invalid.. [[]] []
```

## How to review / test this change

- Log consent at account register (for SMS) or newsletter subscription
- Sync consent
- Check that the consent is synced as expected, and the logs are clear
- Run the Importer, check any failures returned when the import is checked

Related work items: #271648