Merged PR 58314: Add / update try catch when publishing to queues
## What's being changed

In all places where we publish to queues, we are ensuring such calls are wrapped in try/catch statements. I've gone a little beyond this and made sure that all code in critical storefront observers (anything relating to customer interactions, or sales) is wrapped in try/catch.

## Why it's being changed

It's important that observers and plugins in modules must never derail active processes in Magento. The publish call does throw a LocalizedException if the name of the topic is not recognised.

## How to review / test this change

- Smoke test each updated observer, prioritising the ones where errors would impact the storefront
- Pick a couple of critical ones (like OrderSaveAfter) and make them throw an exception by setting a bad topic name in the publisher->publish call

Related work items: #265574