Merged PR 58448: Order place MFTF tests
## What's being changed

This PR adds three new MFTF tests for order placement when this module is enabled. There's a test for a customer order with and without an active Dotdigital account connection, and a test for a subscribed guest order.

## Why it's being changed

Checkout is critical and we want to be sure we haven't accidentally broken it.

## How to review / test this change

- See the tests pass in the build (the whole thing may not be 100% green)
- Run PlaceOrderAsSubscribedGuestTest locally, first changing `Dotdigitalgroup\Email\Model\Subscriber::TOPIC_NEWSLETTER_SUBSCRIPTION` to something bad
- Confirm the test fails

## Notes

Testing of automation enrolment following an order has been pretty thoroughly tested in a new integration test for the OrderSaveAfter observer, so I haven't added automation config / testing here.

I initially had some extra steps in PlaceOrderAsSubscribedGuestTest, to check the contact report and look for a matching subscriber marked as imported. It worked locally, but it refused to in the pipeline. I tried many different things, and ran the MFTF suite against this PR probably about 100 times. In the end I've given up and omitted these steps. I think this is fine because a) our report row filter tests aren't actually very good (they check for the presence of the string 'Imported' in any table cell in the grid) and b) the test still fails if the topic is bad, which is the original scenario we wanted to protect against.

Related work items: #265647