Merged PR 58651: Don't process consent imports as legacy bulk
## What's being changed

Consent imports are not processed alongside legacy file-based sync data.

## Why it's being changed

If consent is included here, the Bulk class looks for an import file. When it doesn't find one it returns null, leading to Error Unknown in the BulkItemPostProcessor.

We could create a new import group for Consent with 'Bulk' and V3 - but these will always fail as well, because of the consent_datetime bug.

So essentially:
- pending 'Bulk' imports for consent won't be picked up
- merchants may have to reset their consent data and re-import it
- in-progress consent imports will work when checked by the V2 checker

Related work items: #273045