Merged PR 58836: Fetch category names snapshot by store
## What's being changed

Product category names are fetched for all categories up front, in the exporter - and not on-demand, per-customer, in a ContactData class.

I've also added a database profiler output to the dotdigital:sync CLI if `--mode=debug` is set.

## Why it's being changed

This approach achieves a significant reduction in database queries for customer sync, mainly reducing SELECT queries on catalog_category_entity_varchar.

## How to review / test this change

- You need a website with at least 2 stores, with customers with orders for each of these stores
- In the category admin, change the names of some popular categories to be different in store B from store A
- Reset customers and run customer sync
- Observe that the store B customers have the expected store-B-specific category names exported

## Notes

The actual reduction in our local tests was -10% for a customer sync of 35 customers across 2 websites. The reduction will be greater the higher the proportion of customers with orders, and the higher the number of categories attached to products in merchant stores.

With the profiler output in the CLI you can do fx:
```
bin/magento dotdigital:sync Customer --mode=debug > output.log
```
and see a full profiler output.

Related work items: #259521