Merged PR 58769: Add config defaults for image types
## What's being changed

Image types configurations should have default values in config.xml. This PR also fixes a bug with abandoned browse for simple products that are part of a bundle, but show the parent (bundle) thumbnail in last browsed data for the simple product.

## Why it's being changed

These keys should properly default to 0.

For abandoned browse, we've intended to use only the id to source an image from the cache (because, since it's WBT, there will always be one available). However, in the ParentFinder, supplying a NULL $imageRole actually meant that the parent's image was being sourced, even if one was available for the child.

## How to review / test this change

- Clear any 'image_types' configs from core_config_data
- Clear cache
- Test setting / saving image type configs
- Reset everything back to default
- Add a WBT profile id
- Visit a product page appending ?dm_i=0,0,0,0,0
- Confirm that the image in the payload is the expected one belonging to the product in view
- Now go to a simple product page that is also part of a bundle (fx Sprite Foam Roller)
- The image in the payload should be the simple product's thumbnail, not the bundle's

Related work items: #249791, #273508