Merged PR 59501: Adjust cron group configs
## What's being changed

The default cron group configurations. All groups now schedule ahead for 10 minutes, and `ddg_automation_main` and `ddg_automation_subs` have an extended schedule lifetime of 10 minutes.

## Why it's being changed

Adjusting the settings in this way is intended to accommodate merchants who do not run cron every minute. Some merchants run cron every 5 minutes or more, and in this case, with our previous configuration, it was typical for scheduled jobs to be missed, and for there to be delays in syncing and importing.

By scheduling jobs further ahead, we ensure that when cron runs it still finds future jobs to process. This as a problem with only scheduling ahead for 4 minutes; it could mean no jobs were scheduled when cron ran. If cron is run even less often than every 5 minutes, this setting can be adjusted in Configuration > Advanced > System.

By extending the schedule lifetime, we reduce the rate of missed jobs for jobs that run less often. If there is a job scheduled to run every 60 mins, even a 5-minute cron will pick it up because it will never be older than 10 minutes.

## How to review / test this change

- Set cron to run on a random 5-minute interval i.e.
```
1-59/5 * * * * /usr/local/bin/php /www/bin/magento cron:run 2>&1 | grep -v "Ran jobs by schedule" >> /www/var/log/magento.cron.log
```
- Set some jobs to run every minute, others every 5, others every 15, others every 60
- Let cron run and check the cron report at intervals
- Note the pattern of missed jobs
- Confirm that jobs scheduled every 60 minutes are still run close to schedule
- Confirm that jobs scheduled every minute are still run at least every 5 minutes

## Notes

- Missed jobs are only problematic for crons scheduled less often (every 15 mins+). If customer sync is set to run every minute, but cron only runs every 5, we will still see missed jobs. This doesn't matter - customer sync will now effectively run on a 5-minute schedule. As long as one job runs every 5 minutes, it will consume all available work anyway.
- The Abandoned Cart cron group does not have an extended schedule lifetime. It should be seen as helpful for merchants to observe missed AC crons because this clearly indicates that some dropped carts are not being picked up. AC cron picks up modified quotes in a very narrow window, so if possible, the `ddg_automation_ac` group should always be run on a 1-minute cron to maximise the number of cart abandonments that are captured.

Related work items: #253510