Merged PR 59401: UpdateContactsWithSubscriberCustomers is not scope-aware
## What's being changed

This fixes a malfunction with the UpdateContactsWithSubscriberCustomers migration when the same email address is connected to multiple websites, but is not subscribed in all scopes.

## Why it's being changed

Currently, if you have a customer in website 1 and a subscriber in website 2 (with the same email address), dotdigital:migrate effectively subscribes you in both scopes.

## How to review / test this change

- Register as a customer on website 1 (do not subscribe)
- Subscribe with the same email address on website 2
- Check that you have 2 rows in email_contact (screenshot the table to refer back to)
- Run dotdigital:migrate
- You should still have 2 rows, but the state of is_subscriber should reflect the correct subscription status per the newsletter_subscriber table

## Notes

This means AbstractUpdater is now not used - it should be deprecated.

Related work items: #275945