Merged PR 59479: Fix single contact deletes if contact is suppressed
## What's being changed

Contact deletions queued with the importer.

## Why it's being changed

A return type was added to the process method in 4.19.0, but this did not account for what happens when a contact is suppressed and we attempt to delete them. Our rather clumsy approach here involves creating the contact (in order to check they exist), then, if they exist, delete them. However, if a contact is suppressed, `$result` is a string ('Contact is suppressed').

It is expected that suppressed contacts cannot be deleted (they need to keep existing to prevent them being added back in the future).

## How to review / test this change

- Find a customer in your system
- Suppress them in DD
- Delete the customer from the Magento admin
- See a deletion queued in the importer table
- See the item processed without error

## Notes

This is a minimal change because we will be moving this deletion to a message queue next year.

Related work items: #276140