Merged PR 61818: Add query to set last_increment_id from last order ids
## What's being changed

The SalesDataManager. A new query is added to fetch the increment_id values of the entity ids from `last_order_id`. Fixes #637.

## Why it's being changed

`last_increment_id` is VARCHAR, therefore MAX will not reliably fetch the last increment id (which really is the increment id of the most recent order id). Furthermore:
- Magento by default will use the store id to differentiate increment ids e.g. store 3 increment ids start at   300000001, so they will automatically be picked by MAX over store 1 orders by the same customer.
- Merchants can configure their own custom order increment systems e.g. ORD-0001

Example:
| entity_id | increment_id |
|--|--|
| 22	| 3000000013 |
| 23	| 3000000014 |
| 24	| 3000000015 |
| 27	| 3000000018 |
| 28	| 3000000019 |
| 36	| 000000007 |

The correct last_increment_id for this customer would be 000000007, not 3000000019 (as selected before this change).

In terms of implementation - yes, it's another query BUT the alternatives were:
- subquery in `fetchBaseOrderCollection`
- add increment_id to the `fetchOrderedProductIds` query, which also uses last_order_id (in one of its invocations)

## How to review / test this change

- On a multi-store Magento, place an order for the same customer on 2 different stores
- Place the order for store 1 (or the default store) last
- Map all data fields
- Run customer sync
- The exported last_increment_id should be the increment_id of the last order

Related work items: #284929