Merged PR 62412: Translate LASTSUBSCRIBED data field for accounts in supported languages
## What's being changed

The LASTSUBSCRIBED data field is now translated prior to being used in the PlatformChangeManager, based on a defined list of 11 languages, as supported by Dotdigital.

Before checking modified contacts, we look up the relevant DD account locale. If this is 'non-en', we try to find a matching translation. If we do, we double check against the account data fields. Then finally we return the translated data field name for use in the API call.

## Why it's being changed

For Dotdigital accounts where the locale is set at account creation, a subset of data fields are translated into the desired account culture. Whereas for customer sync, merchants can map e.g. 'VORNAME' to first name, 'LASTSUBSCRIBED' cannot be mapped and is hard-coded in the PlatformChangeManager.

This means that for affected merchants, the PlatformChangeManager doesn't work, and in fact just logs an error every time the ddg_automation_platform_modified_contacts cron runs.
```
Error when running a cron job: Call to a member function all() on null
```

## How to review / test this change

- Turn off cron
- Without changing your locale, run `bin/magento dotdigital:task PlatformChangeManager`
- The script should run as before for locales containing 'en'
- Change your locale in Dotdigital Account Settings to one of the other locales offered
- Run the script again, it won't find a matching DF so should default back to LASTSUBSCRIBED
- Use an API user in one of your websites for a merchant with translated data fields
- Check the script runs as intended with the translated data field
- Remove the merchant API user from your system

## Notes
Github ref https://github.com/dotmailer/dotmailer-magento2-extension/pull/640

Related work items: #286783