Merged PR 62151: Use tax helper to get tax prices in catalog sync
## What's being changed

This PR aligns and fixes our logic for setting product prices (with and without tax).

## Why it's being changed

Currently we have a scattered approach where Catalog sync, WBT and Hyvä WBT all do slightly different things. We are aligning to a system that uses Magento's TaxHelper to get the tax price. This is what Hyvä has been using for WBT - but there is one key difference. IF Sales > Tax > Catalog Prices is set to 'Including Tax', we add the rate %.

Before this (depending on the context) we might do one of:
```
price = price - rate // this would be price_excl_tax (which we don't sync)
or
price_incl_tax = price + rate // too high if prices should include tax already
```

## How to review / test this change

_Setup_
- Set a tax zone and rate (something simple like +20% everywhere)
- Set Sales > Tax > Catalog Prices to 'Excluding Tax'
- Set a simple product to have price £100, special price £50
- View it on the storefront, you should see £55 (reduced from £120)

_Simple_
- Run catalog sync, you should see price 100, price_incl_tax 120, special price 50, special_price_incl_tax 55
- Set Sales > Tax > Catalog Prices to 'Including Tax'
- Mark your simple product for re-sync
- Run catalog sync, you should see price 100, price_incl_tax 100, special price 50, special_price_incl_tax 50

_Other product types_
- You should now repeat simple and test price export for configurable, group and bundle products.
- We have business logic here to present the minimum price of each (which hasn't been changed). Again, it's easier if you set e.g. the lowest simple child price of a configurable parent to £20, special £10, so the price points are super-clear.

## Notes

What we're trying to do here is e.g.
- merchant has product priced at £100 including tax
- merchant charges VAT but this is never publicised, it's always included in the price
- price should be £100, price_incl_tax should be £100

If catalog sync sets the price_incl_tax to £120, that's incorrect, and equally setting the price to £80 is also incorrect, because that £80 price point is actually price_excl_tax (which we don't sync).

Related work items: #279645, #285662